+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 ## 4.2.1
 
+- Updated Catalyst Fund6 and Fund7 voting registration dates ([PR 2639](https://github.com/input-output-hk/daedalus/pull/2639))
+
 ## 4.2.0
 
 ### Features